pyproject.toml
[project]
name = "reasonbench"
version = "0.1.0"
description = "LLM Adversarial Reasoning Evaluation System"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0,<3",
    "anthropic>=0.40",
    "scikit-learn>=1.5",
    "requests>=2.31",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
gemini = ["google-generativeai>=0.8"]
all-providers = [
    "reasonbench[openai]",
    "reasonbench[gemini]",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.9",
    "mypy>=1.14",
    "reasonbench[all-providers]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["reasonbench"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["UP042"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
